LocalBiz-Lead-IQ-main/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
tldextract==5.1.2
python-dotenv==1.0.1
fastapi==0.115.8
uvicorn==0.30.6
pydantic==2.10.6
